Add files via upload
This program is meant to take in an array of integers given by the user, print out the array of integers, then determine if the array is in ascending order or not. If the array is not, the program will return a "0". If it is, the program will return a "1".